+++ .formatter.exs
@@ -4,6 +4,7 @@ locals_without_parens = [
   field: 2,
   field: 3,
   relation: 3,
+  version: 1,
   on_create: 2,
   on_update: 2,
   on_delete: 2,